napari-mmc-tools/pyproject.toml
# ===============================
# pyproject.toml
# ===============================
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "napari-mmc-tools"
version = "0.1.0"
description = "Custom tools for napari-micromanager: shared TL launcher and MDA positions editor"
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "napari>=0.4.19",
  "magicgui>=0.7",
  "pymmcore-plus>=0.11",
  "useq-schema>=0.5",
  "PyYAML>=6",
]

[project.urls]
Homepage = "https://example.com/napari-mmc-tools"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"napari_mmc_tools" = ["napari.yaml"]

[project.entry-points."napari.manifest"]
"napari-mmc-tools" = "napari_mmc_tools:napari.yaml"
